+++ org.eclipse.lsp4j/src/main/java/org/eclipse/lsp4j/SignatureHelpTriggerKind.java
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright (c) 2016 TypeFox and others.
+ * Copyright (c) 2016-2018 TypeFox and others.
  * 
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -12,63 +12,30 @@
 package org.eclipse.lsp4j;
 
 /**
- * The kind of a completion entry.
+ * How a signature help was triggered.
+ * 
+ * Since 3.15.0
  */
-public enum CompletionItemKind {
+public enum SignatureHelpTriggerKind {
 	
-	Text(1),
+	/**
+	 * Signature help was invoked manually by the user or by a command.
+	 */
+	Invoked(1),
 	
-	Method(2),
+	/**
+	 * Signature help was triggered by a trigger character.
+	 */
+	TriggerCharacter(2),
 	
-	Function(3),
-	
-	Constructor(4),
-	
-	Field(5),
-	
-	Variable(6),
-	
-	Class(7),
-	
-	Interface(8),
-	
-	Module(9),
-	
-	Property(10),
-	
-	Unit(11),
-	
-	Value(12),
-	
-	Enum(13),
-	
-	Keyword(14),
-	
-	Snippet(15),
-	
-	Color(16),
-	
-	File(17),
-	
-	Reference(18),
-	
-	Folder(19),
-	
-	EnumMember(20),
-	
-	Constant(21),
-	
-	Struct(22),
-	
-	Event(23),
-	
-	Operator(24),
-	
-	TypeParameter(25);
+	/**
+	 * Signature help was triggered by the cursor moving or by the document content changing.
+	 */
+	ContentChange(3);
 	
 	private final int value;
 	
-	CompletionItemKind(int value) {
+	SignatureHelpTriggerKind(int value) {
 		this.value = value;
 	}
 	
@@ -76,8 +43,8 @@ public enum CompletionItemKind {
 		return value;
 	}
 	
-	public static CompletionItemKind forValue(int value) {
-		CompletionItemKind[] allValues = CompletionItemKind.values();
+	public static SignatureHelpTriggerKind forValue(int value) {
+		SignatureHelpTriggerKind[] allValues = SignatureHelpTriggerKind.values();
 		if (value < 1 || value > allValues.length)
 			throw new IllegalArgumentException("Illegal enum value: " + value);
 		return allValues[value - 1];