+++ org.eclipse.lsp4j/src/main/xtend-gen/org/eclipse/lsp4j/TextDocumentClientCapabilities.java
@@ -181,7 +181,6 @@ public class TextDocumentClientCapabilities {
   /**
    * Capabilities specific to `textDocument/selectionRange` requests
    */
-  @Beta
   private SelectionRangeCapabilities selectionRange;
   
   @Pure